refactor: Simplify dataset loading and preprocessing
- Remove text column specification from dataset loading
- Update argument parsing for train and eval datasets to support multiple datasets
- Add isort configuration in pyproject.toml to match Black formatting
