+++ pyproject.toml
@@ -1,2 +1,5 @@
 [tool.black]
 line-length = 120
+
+[tool.isort]
+profile = "black"
